Increase size of user_agent column
If a user_agent string is is longer than the Laravel default value of the string field, inserting the audit causes an exception which in turn breaks the application. Changing the user_agent string to always be 2048 characters will solve this issue. While there is no limit on the size of a user agent, somewhere between 512 - 2048 seems to be the recommended max size.